Improving setFavorite

+++ lib/favorite/favorite_service.dart
@@ -5,16 +5,18 @@ import 'package:car_project/favorite/favorite_dao.dart';
 class FavoriteService {
   static FavoriteDAO _favoriteDAO = FavoriteDAO();
 
-  static setFavorite(Car car) {
-    Favorite favorite = Favorite(car.id, car.nome);
-    _favoriteDAO.save(favorite);
+  static setFavorite(Car car) async {
+    final isFavorite = await _favoriteDAO.isFavorite(car.id);
+
+    if (isFavorite) {
+      _favoriteDAO.delete(car.id);
+    } else {
+      Favorite favorite = Favorite(car.id, car.nome);
+      _favoriteDAO.save(favorite);
+    }
   }
 
   static Future<bool> isFavorite(Car car) async {
     return _favoriteDAO.isFavorite(car.id);
-  }
-
-  static void removeFavorite(Car car) {
-    _favoriteDAO.delete(car.id);
   }
 }
